grammar: resolve type/value parsing ambiguities
The upstream railroad diagram has ambiguities around the way values
should be parsed: numeric types are inherently ambiguous (e.g. '1' is
both a valid integer and float), and thus values can only be
semantically parsed correctly provided we know the proper type for the
structure.

Luckily, this is easy to accomplish: we can simply abstract out the rule
generation logic, and make a 'choice' for parsers -- based on whether a
type parser, succeeds, we simply choose the right value parser.

This not only cleans up the parsing behavior, but actually makes the
grammar description itself more readable.

This also fixes a bug where 'type' nodes for a structure were
accidentally hidden, when they shouldn't have been.

Update tests to cope with everything.

Signed-off-by: Austin Seipp <aseipp@pobox.com>